# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gotenberg
  version: 7.x
  license:
    name: MIT
    url: 'https://github.com/gotenberg/gotenberg/blob/main/LICENSE'
  contact:
    url: 'https://github.com/gotenberg/gotenberg'
  description: >-
    A Docker-powered stateless API for PDF files.
...
servers:
  - url: 'http://localhost:3000'
    description: Local server with the default Docker image and port
...
paths:
...
  /forms/libreoffice/convert:
    post:
      tags:
        - libreoffice
      summary: Convert an Office document to PDF or HTML
      externalDocs:
        url: https://gotenberg.dev/docs/modules/libreoffice
      description: >-
        This route accepts multipart/form-data requests and files with the following extensions:

        .bib  .doc  .xml  .docx  .fodt  .html  .ltx  .txt  .odt  .ott  .pdb  .pdf  .psw  .rtf
        .sdw  .stw  .sxw  .uot  .vor  .wps  .epub  .png  .bmp  .emf  .eps  .fodg  .gif  .jpg
        .met  .odd  .otg  .pbm  .pct  .pgm  .ppm  .ras  .std  .svg  .svm  .swf  .sxd  .sxw
        .tiff  .xhtml  .xpm  .fodp  .potm  .pot  .pptx  .pps  .ppt  .pwp  .sda  .sdd  .sti
        .sxi  .uop  .wmf  .csv  .dbf  .dif  .fods  .ods  .ots  .pxl  .sdc  .slk  .stc  .sxc
        .uos  .xls  .xlt  .xlsx  .tif  .jpeg  .odp

        By default, if you send more than one file to convert, the route returns a ZIP archive of the
        resulting PDF files. However, you may prefer to merge all the PDF files into an individual PDF file.

        > **Attention:** The files will be merged alphabetically for the
        resulting PDF.

        You may also specify the page ranges to convert from the incoming Office
        documents. The expected format is the same as the one from the print
        options of LibreOffice, e.g. `1-1` or `1-4`.

        > **Attention:** if more than one document, the page ranges will be
        applied for each document.

        You can also request conversion to HTML format, but this disables the merge capabilities.

        See externalDocs for more details.
      parameters:
        - in: header
          name: Gotenberg-Output-Filename
          description: >-
            By default, the API generates a UUID filename.
            However, you may also specify the filename per request,
            thanks to the Gotenberg-Output-Filename header.
            Caution! The API adds the file extension automatically; you don't have to set it.
          schema:
            type: string
          required: false
        - in: header
          name: Gotenberg-Trace
          description: >-
            The trace, or request ID, identifies a request in the logs.

            By default, the API generates a UUID trace for each request.
            However, you may also specify the trace per request, thanks to the Gotenberg-Trace header.
          schema:
            type: string
          required: false
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OfficeConvertRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulConvert'
        '400':
          description: Bad Request, e.g. Both 'pdfFormat' and 'nativePdfA1aFormat' form values are provided
...
components:
  schemas:
...
    OfficeConvertRequestBody:
      title: Office Conversion Request Body
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        nativePageRanges:
          type: string
          example: 1-4
          description: >-
            The page ranges to be converted to PDF for the incoming Office
            documents. **If there are multiple files sent to the API, this page
            range will apply to all of the documents**. Empty means all pages.
        nativePdfA1aFormat:
          type: boolean
          description: >-
            Use unoconv to convert the resulting PDF to the 'PDF/A-1a' format.
            Caution! You cannot use both nativePdfA1aFormat and pdfFormat form fields.
        pdfFormat:
          type: string
          description: >-
            The PDF format of the resulting PDF.
            Caution! You cannot use both nativePdfA1aFormat and pdfFormat form fields.
          example: PDF/A-1a
        landscape:
          type: boolean
          example: true
          default: false
          description: >-
            The default orientation for rendering the page is "portrait" mode.
            By sending "landscape" parameter, you can ask the output to be
            landscape.
        merge:
          type: boolean
          description: >-
            Merge all PDF files into an individual PDF file.
        htmlFormat:
          type: boolean
          description: >-
            Convert the input document to HTML, rather than PDF.  Caution!
            You cannot use with nativePdfA1Format, pdfFormat, nativePageRanges, or
            merge options!
        importFormat:
          type: string
          example: text
          description: >-
            Explicitly set the import format for the document, which can be useful
            when converting files without an extension.
        importFilters:
          type: string
          example: FilterOptions=SHIFT_JIS
          description: >-
            Set import filters, which can be particularly useful in converting
            input files created with a different character encoding.
      required:
        - files
...
  securitySchemes: { }
  responses:
...
    SuccessfulConvert:
      description: Resulting document from the conversion.
      content:
        application/pdf:
          schema:
            type: string
            format: binary
        text/html:
          schema:
            type: string